groups/serializers.py
from rest_framework import serializers
from .models import Category, Link


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ('id', 'name', 'image')


class LinkSerializer(serializers.ModelSerializer):
    class Meta:
        model = Link
        fields = ('id', 'category', 'name', 'link', 'details')

