requirements.txt
bootstrap-admin==0.3.9
Django==2.1
djangorestframework==3.8.2
Pillow==5.2.0
pytz==2018.5
